Refine legacy reader context and document guidelines
